+++ test/Unit/Attribute/MaximumDurationTest.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2021-2023 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/phpunit-slow-test-detector
+ */
+
+namespace Ergebnis\PHPUnit\SlowTestDetector\Test\Unit\Attribute;
+
+use Ergebnis\DataProvider;
+use Ergebnis\PHPUnit\SlowTestDetector\Attribute;
+use Ergebnis\PHPUnit\SlowTestDetector\Duration;
+use Ergebnis\PHPUnit\SlowTestDetector\Exception;
+use Ergebnis\PHPUnit\SlowTestDetector\Test;
+use PHPUnit\Framework;
+
+#[Framework\Attributes\CoversClass(Attribute\MaximumDuration::class)]
+final class MaximumDurationTest extends Framework\TestCase
+{
+    use Test\Util\Helper;
+
+    #[Framework\Attributes\DataProviderExternal(DataProvider\IntProvider::class, 'lessThanZero')]
+    #[Framework\Attributes\DataProviderExternal(DataProvider\IntProvider::class, 'zero')]
+    public function testConstructorRejectsInvalidValue(int $milliseconds): void
+    {
+        $this->expectException(Exception\InvalidMilliseconds::class);
+
+        Duration::fromMilliseconds($milliseconds);
+    }
+
+    public function testConstructorSetsValue(): void
+    {
+        $milliseconds = self::faker()->numberBetween(1);
+
+        $maximumDuration = new Attribute\MaximumDuration($milliseconds);
+
+        self::assertSame($milliseconds, $maximumDuration->milliseconds());
+    }
+}